Add files via upload
Market-Analysis-of-Hollywood-Movies
The dataset contains information about the top 1000 highest grossing Hollywood films. It is up to date as of 10th January 2022. The purpose is to identify patterns, trends, and useful data from the dataset for Hollywood movies. In order to achieve this, Advanced Excel is used for assessing and analyzing the data